refactor: helpers folder
