+++ __tests__/unit_tests/custom-date.test.ts
@@ -2,7 +2,7 @@ import {
   determine_next_day_function,
   determine_next_execute_date_freq,
   getOffsetDay,
-} from "../../utils/helpers/custom-date-fns.ts";
+} from "../../utils/helpers/date/custom-date-fns.ts";
 
 let today: Date;
