+++ src/Events/app_home.ts
@@ -3,7 +3,7 @@ import app from "../../utils/config/slack-config.ts";
 import { AXIOM_DATA_SET } from "../../utils/constants/consts.ts";
 
 import Sentry from "../../utils/config/sentry.config.ts";
-import { generate_newsletter } from "../../utils/helpers/generate-newsletter.ts";
+import { generate_newsletter } from "../../utils/helpers/misc/generate-newsletter.ts";
 
 const transaction = Sentry.startTransaction({
   op: "app_home",
