+++ src/Events/greetings.ts
@@ -3,7 +3,7 @@
 
 import Axiom from "../../utils/config/axiom-config.ts";
 import app from "../../utils/config/slack-config.ts";
-import { generate_welcome_message } from "../../utils/helpers/generate_message.ts";
+import { generate_welcome_message } from "../../utils/helpers/misc/generate_message.ts";
 
 import { AXIOM_DATA_SET } from "../../utils/constants/consts.ts";
