+++ src/Events/post_newsletter.ts
@@ -1,5 +1,5 @@
 import app from "../../utils/config/slack-config.ts";
-import { generate_newsletter_post } from "../../utils/helpers/generate-newsletter.ts";
+import { generate_newsletter_post } from "../../utils/helpers/misc/generate-newsletter.ts";
 
 import Axiom from "../../utils/config/axiom-config.ts";
 import { AXIOM_DATA_SET } from "../../utils/constants/consts.ts";
