+++ src/Events/send-scheduled-message.ts
@@ -1,6 +1,6 @@
 import app from "../../utils/config/slack-config.ts";
-import { generate_scheduled_messages } from "../../utils/helpers/generate-scheduled-messages.ts";
-import { prep_final_scheduled_message } from "../../utils/helpers/prep-final-scheduled-message.ts";
+import { generate_scheduled_messages } from "../../utils/helpers/schedule_message/generate-scheduled-messages.ts";
+import { prep_final_scheduled_message } from "../../utils/helpers/schedule_message/prep-final-scheduled-message.ts";
 import type { day } from "../../utils/types/projectTypes.ts";
 
 import Axiom from "../../utils/config/axiom-config.ts";
