+++ utils/helpers/general/obj-has-property.ts
