+++ utils/helpers/misc/search-user-helper.ts
