+++ utils/helpers/schedule_message/check-if-message-scheduled.ts
@@ -1,5 +1,5 @@
-import { MATCH_GREETING_MESSAGE } from "../constants/consts.ts";
-import type { ScheduleMessage } from "../types/projectTypes.ts";
+import { MATCH_GREETING_MESSAGE } from "../../constants/consts.ts";
+import type { ScheduleMessage } from "../../types/projectTypes.ts";
 
 export const check_if_message_is_scheduled = (
   msgs_to_be_scheduled: ScheduleMessage[],
